Add transformation rules by default for the available license bundles
Often a license name is defined properly, but the url is slightly different (e.g. http instead of https),
but defining a rule is then quite annoying. Therefor, the the normalizer defined its own rules by just
using all the license-bundles and create a name-rule for every bundle where the name-pattern is set to the
license-name of the licence-bundle. This will then normalize all licenses with a valid name, but a
different url. The same is done for the url. So all licenses with the same url, but a differnet name
are also normalized.