Allow to configure if name or url of bundle whould be transformed
Sometimes, libraries have e.g. an Apache2 License but the license-url points to their own license
containing proper company information instead of the placeholders in the central license-templates.
In those cases, the url shouldn't be tranformed, but only the name. This is now added. On the roles,
it can be specified what to transform in case the rule matches.