Add support for checking if all dependency licenses are allowed
When having a project, not always all licenses are allowed to be used
like GPL in commercial programs. This commit allows to define a file
where allowed licenses can be specified and by running the new task
'checkLicense', all the dependencies are checked against that file.
If licenses are not allowed, the task fails.

fixing LicenseChecker bugs

delete ChcekLicenseTaskSpec buildScript

adding moduleVersion to LicenseCheckr
add equal comparison to LicenseChecker