Better license mapping from manifest and via transformations
Some old artifacts have very little to no license information, so such dependencies
are mapped using a license-transformation rule using a `modulePattern`.
Although the pattern matches and the license bundle has the correct URL, the dependency
appears in the "Unknown" category in the `InventoryHtmlReportRenderer`, which is then
incorrect.

Example "bad" dependency, that needs such a mapping is `javax.servlet.jsp:jsp-api:2.1`.

* InventoryHtmlReportRenderer: slightly enlarge the left navigation bar
* ManifestData: add new `licenseUrl` field, that's filled when reading the manifest, not interpreted when needed
* LicenseBundleNormalizer: more debug messages