Fix deprecation warnings when running with Gradle 6.0
Messages look like:

```
Property 'classpath' is declared without normalization specified. Properties of cacheable work must declare their normalization via @PathSensitive, @Classpath or @CompileClasspath. Defaulting to PathSensitivity.ABSOLUTE. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
Property 'config.filtersCache' is private and annotated with @Nested. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
Property 'config.importersCache' is private and annotated with @Nested. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
Property 'config.licenseReportExtensionSnapshot' is private and annotated with @Input. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
Property 'config.renderersCache' is private and annotated with @Nested. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
Property 'config.renderer' is not annotated with an input or output annotation. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
Property 'config.filtersCache.$0.filterConfigForCache' is private and annotated with @Input. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
Property 'config.filtersCache.$0.bundleMap' is not annotated with an input or output annotation. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
Property 'config.filtersCache.$0.config' is not annotated with an input or output annotation. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
Property 'config.filtersCache.$0.duplicateFilter' is not annotated with an input or output annotation. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
Property 'config.filtersCache.$0.normalizerConfig' is not annotated with an input or output annotation. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
Property 'config.renderersCache.$0.fileNameCache' is private and annotated with @Input. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
Property 'config.renderersCache.$1.fileNameCache' is private and annotated with @Input. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
Property 'config.renderersCache.$2.fileNameCache' is private and annotated with @Input. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
Property 'config.renderersCache.$2.onlyOneLicensePerModuleCahce' is private and annotated with @Input. This behaviour has been deprecated and is scheduled to be removed in Gradle 7.0.
```