Partial ix for sub-projects in filter
Updates the ExcludeTransitiveDependenciesFilter so that it can find
configurations correctly by searching the project tree rather than
just looking at the root project.
Note this is a partial fix and assumes that the configuration(s)
specified in the plugin config are unique across sub-projects.
The more complete fix would be to merge configurations with the
same name (see ProjectData.groovy).