fix missing override for InventoryReportRenderer
After introducing the delayed loading of the overrides file with
PR #286, the overrides for the first dependency in the
InventoryReportRenderer were not applied.

It is not entirely clear to me, why this happens. I just assume it has
to do with groovy-magic related to automatically calling the getters of
properties.

To resolve this, the base-property was renamed, so it doesn't conflict
with the getter.