feat(pnpm): add Pnpm licenses importer
The license-checker module doesn't work with pnpm unfortunately so it
cannot be used.

Pnpm has it's own built-in license extraction. Add an importer similar
to the npx importer from pnpm's `pnpm licenses list` command.

`pnpm licenses list` doesn't download the licenses, nor supply the url,
but the URL appears to be filled in later.

The licenses are not downloaded so there is no directory structure when
compared to the Java dependencies.