Track allowedLicensesFile as task input file
Use `@InputFile` rather than `@Input` so that we track file contents,
not just the path. This fixes up-to-date checking and build caching for
the checkLicense task when the contents of the allowed license file
changes.

Also, use RELATIVE sensitivity instead of the default ABSOLUTE for
a bit more flexibility in how this path is evaluated.

Co-Authored-By: Jon Parise <jon@indelible.org>